blocking or non-blocking logger, gated by features
