+++ Cargo.toml
@@ -3,5 +3,9 @@ name = "traccia"
 version = "0.1.0"
 edition = "2024"
 
+
 [dependencies]
 dyn-clone = "1.0.18"
+
+[features]
+blocking = []
