+++ src/impl/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(not(feature = "blocking"))]
+pub mod r#async;
+
+#[cfg(feature = "blocking")]
+pub mod blocking;
